started writing docs for compute graphs
